moving Enum classes into enums.py
