+++ python/enums.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+
+class ProductUnit(Enum):
+    EACH = 1
+    KILO = 2
+
+
+class SpecialOfferType(Enum):
+    THREE_FOR_TWO = 1
+    TEN_PERCENT_DISCOUNT = 2
+    TWO_FOR_AMOUNT = 3
+    FIVE_FOR_AMOUNT = 4
