+++ python/model_objects.py
@@ -1,6 +1,3 @@
-from enum import Enum
-
-
 class Product:
     def __init__(self, name, unit):
         self.name = name
@@ -11,18 +8,6 @@ class ProductQuantity:
     def __init__(self, product, quantity):
         self.product = product
         self.quantity = quantity
-
-
-class ProductUnit(Enum):
-    EACH = 1
-    KILO = 2
-
-
-class SpecialOfferType(Enum):
-    THREE_FOR_TWO = 1
-    TEN_PERCENT_DISCOUNT = 2
-    TWO_FOR_AMOUNT = 3
-    FIVE_FOR_AMOUNT = 4
 
 
 class Offer:
