+++ python/tests/test_supermarket.py
@@ -2,10 +2,11 @@ import unittest
 
 from approvaltests import verify
 
-from model_objects import Product, SpecialOfferType, ProductUnit, ShoppingCart
+from model_objects import Product, ShoppingCart
 from receipt_printer import ReceiptPrinter
 from teller import Teller
 from tests.fake_catalog import FakeCatalog
+from enums import ProductUnit, SpecialOfferType
 
 
 class SupermarketTest(unittest.TestCase):
